+++ src/plot/layer/geom/rule.rs
@@ -29,6 +29,7 @@ impl GeomTrait for Rule {
                 ("linewidth", DefaultAestheticValue::Number(1.0)),
                 ("opacity", DefaultAestheticValue::Number(1.0)),
                 ("linetype", DefaultAestheticValue::String("solid")),
+                ("label", DefaultAestheticValue::Null),
             ],
         }
     }